move events into a separate Go file
